Move Streams/StreamK* to Internal/Data/Stream
