+++ benchmark/StreamKOps.hs
@@ -19,7 +19,7 @@ import Prelude
 import qualified Prelude as P
 import qualified Data.List as List
 
-import qualified Streamly.Streams.StreamK as S
+import qualified Streamly.Internal.Data.Stream.StreamK as S
 import qualified Streamly.Streams.Prelude as SP
 import qualified Streamly.Internal.Data.SVar as S
