+++ src/Streamly/Internal/Memory/ArrayStream.hs
@@ -51,7 +51,7 @@ import Prelude hiding (length, null, last, map, (!!), read, concat)
 
 import Streamly.Internal.Memory.Array.Types (Array(..), length)
 import Streamly.Streams.Serial (SerialT)
-import Streamly.Streams.StreamK.Type (IsStream)
+import Streamly.Internal.Data.Stream.StreamK.Type (IsStream)
 
 import qualified Streamly.Internal.Memory.Array as A
 import qualified Streamly.Internal.Memory.Array.Types as A
