+++ src/Streamly/Streams/Serial.hs
@@ -67,10 +67,10 @@ import Text.Read (Lexeme(Ident), lexP, parens, prec, readPrec, readListPrec,
                   readListPrecDefault)
 import Prelude hiding (map, mapM)
 
-import Streamly.Streams.StreamK (IsStream(..), adapt, Stream, mkStream,
+import Streamly.Internal.Data.Stream.StreamK (IsStream(..), adapt, Stream, mkStream,
                                  foldStream)
 import qualified Streamly.Streams.Prelude as P
-import qualified Streamly.Streams.StreamK as K
+import qualified Streamly.Internal.Data.Stream.StreamK as K
 import qualified Streamly.Internal.Data.Stream.StreamD as D
 
 #include "Instances.hs"
